requirements.txt
confluent-kafka
python-dotenv
psycopg2-binary
opensearch-py